djangoprojects/backendwithjwt/requirements.txt
Django~=2.0.0
channels~=2.0,>=2.0.2
channels_redis~=2.0
djangorestframework~=3.9.0
coreapi~=2.3.3
django-cors-headers~=2.4.0
djangorestframework-simplejwt~=3.3